Refactor integration scripts to use packages
Since we export the generated solidity contract factories now,
we can make use of nice import paths and versioning for our integration
scripts.
