--- integration-scripts/contracts/Oracle.sol
@@ -1,3 +1,0 @@
-pragma solidity 0.4.24;
-
-import "../../evm/contracts/Oracle.sol";
