+++ integration-scripts/contracts/RunLog.sol
@@ -1,6 +1,6 @@
-pragma solidity ^0.4.23;
+pragma solidity ^0.4.24;
 
-import "../../evm/contracts/Chainlinked.sol";
+import "chainlink/contracts/Chainlinked.sol";
 
 contract RunLog is Chainlinked {
   uint256 constant private ORACLE_PAYMENT = 1 * LINK; // solium-disable-line zeppelin/no-arithmetic-operations
