+++ integration-scripts/package.json
@@ -7,11 +7,9 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "generate-typings": "echo \"\u001b[1;33mWARN: Please import the generated contract factory classes for LinkTokenV05, Coordinator and MeanAggregator directly from the chainlinkv0.5 package when exported.\u001b[0m\" && typechain --target ethers --outDir src/generated \"{build/contracts/*,../evm/v0.5/test/support/LinkTokenV05.json,../evm/v0.5/build/contracts/Coordinator.json,../evm/v0.5/build/contracts/MeanAggregator.json}\"",
-    "build:contracts": "truffle compile",
-    "postbuild:contracts": "yarn generate-typings",
-    "prebuild": "yarn build:contracts",
-    "build": "tsc",
+    "generate-typings": "typechain --target ethers --outDir src/generated \"dist/artifacts/*\"",
+    "postgenerate-typings": "yarn workspace chainlink export-typings ../integration-scripts/src/generated ../integration-scripts/dist/src/generated",
+    "build": "sol-compiler && yarn generate-typings && tsc",
     "setup": "yarn build",
     "format": "prettier --write \"src/**/*\"",
     "lint": "eslint --ext .ts src",
@@ -25,6 +23,22 @@
     "initiate-service-agreement": "node ./dist/initiateServiceAgreement",
     "start-echo-server": "node ./dist/echoServer"
   },
+  "dependencies": {
+    "@0x/sol-compiler": "^3.1.15",
+    "@0x/sol-trace": "^2.0.20",
+    "@chainlink/eslint-config": "0.0.1",
+    "@chainlink/prettier-config": "^0.0.1",
+    "body-parser": "^1.18.3",
+    "chainlink": "0.7.8",
+    "chainlinkv0.5": "0.0.2",
+    "chalk": "^2.4.2",
+    "ethers": "^4.0.37",
+    "express": "^4.16.4",
+    "link_token": "^1.0.6",
+    "request-promise": "4.2.4",
+    "shelljs": "^0.8.3",
+    "source-map-support": "^0.5.13"
+  },
   "devDependencies": {
     "@chainlink/eslint-config": "0.0.1",
     "@chainlink/prettier-config": "^0.0.1",
@@ -33,25 +47,10 @@
     "@types/shelljs": "^0.8.5",
     "debug": "4.1.1",
     "eslint": "6.3.0",
-    "typechain": "1.0.1",
-    "typechain-target-ethers": "^1.0.0-beta.1",
-    "typescript": "^3.6.3"
-  },
-  "dependencies": {
-    "@0x/sol-compiler": "^3.1.15",
-    "@0x/sol-trace": "^2.0.20",
-    "@chainlink/eslint-config": "0.0.1",
-    "@chainlink/prettier-config": "^0.0.1",
-    "body-parser": "^1.18.3",
-    "chainlink": "0.7.7",
-    "chalk": "^2.4.2",
-    "ethers": "^4.0.37",
-    "express": "^4.16.4",
-    "link_token": "^1.0.6",
-    "request-promise": "4.2.4",
-    "shelljs": "^0.8.3",
-    "source-map-support": "^0.5.13",
-    "truffle": "^5.0.39"
+    "typechain": "1.0.3",
+    "typechain-target-ethers": "^1.0.1",
+    "typescript": "^3.6.3",
+    "@0x/sol-compiler": "^3.1.15"
   },
   "prettier": "@chainlink/prettier-config"
 }
